requirements.txt
customtkinter
httpx
Pillow